With very long file names, VS 2010 was unable to compile files.
At some point in the past VS 2010 failed some tests with custom commands when
relative paths were not used.  It seems that those problems have been fixed.
However, the relative paths apparently are appended to the current working
directoy before vs accesses the file. So, with a long path, relative paths
cause it to create a combined path that is too long.