IRIX: Work around trailing separator removal in RPATH (#12173)
At least one version of "ld" on this platform removes one trailing
separator from the end of the -rpath value.  This interferes with
CMake's RPATH placeholder generation needed to rewrite the RPATH during
installation.  Work around the problem by appending one '/' to the end
of the build-tree copy of the RPATH so it will not be truncated.