Fix for bug #11927, external project git clone step always runs vs10.
In cmMakefile.cxx GetCMakeCFGInitDirectory is replaced with
GetCMakeFilesDirectory for .rule files.  In some cases with
external projects, that directory will not exist.  With vs10
the .rule files must exist or the rule will run with every build.
This fix creates the path that the .rule file is in.  In addition,
it is now a CMake error if the .rule file can not be created.