find_library: Use lib->lib64 conversion in CXX-only projects (#12247,#12248)
Do not require the C language to be enabled to do lib->lib64 conversion.
The check was originally added by commit a5825cd1 (check in new find
stuff, 2006-03-02) to ensure that CMAKE_SIZEOF_VOID_P is set.  Since
commit 3fdf1411 (FIND_LIBRARY should not require CMAKE_SIZEOF_VOID_P,
2008-02-11) the code following the check does not fail when the variable
is not set, so just remove the original check.  This allows conversion
for any language that is enabled so long as the address size is known.

Reported-by: Christoph Höger <choeger@cs.tu-berlin.de>