Suppress Qt warning for dashmacmini5 builds
Qt 4.7 and earlier produce a "This version of Mac OS X is unsupported"
warning on Lion, even though they work fine once built. We'll upgrade
this machine to use Qt 4.8 when it's officially released. Until then,
suppress this warning so that we don't miss other "more real" warnings
on the dashboard.