Use version 11.0 for 12.x and 9.10 for 10.x intel versions to fix 12.1 vsIDE.
The intel compiler for 12.0 and 12.1 are known to expect the file version
to be 11.0 in the .vfproj file.  For 10.x it should be 9.10.  Prior to
this fix 12.0 and 10.1 were the only values checked. If those did not match
the actual version of intel was put in the vfproj file causing an error
about future version load attempt in the IDE.